Tools/asset-pipeline/requirements.txt
requests>=2.31.0
pymongo>=4.6.0
pymilvus>=2.3.0
boto3>=1.34.0
pandas>=2.1.0
Pillow>=10.1.0
tqdm>=4.66.0
